@@ -42,17 +42,20 @@
     "pouchdb-adapter-leveldb": "^6.1.0",
     "pouchdb-core": "^6.1.0",
     "pouchdb-quick-search": "^1.2.0",
-    "react": "15.1.0",
+    "react": "^15.4.2",
     "react-addons-css-transition-group": "^15.5.2",
     "react-addons-shallow-compare": "^15.5.2",
     "react-bootstrap": "^0.30.8",
     "react-bootstrap-grid": "^1.0.0",
-    "react-dom": "15.1.0",
-    "react-selectize": "^2.1.0"
+    "react-dom": "^15.4.2",
+    "react-selectize": "^2.1.0",
+    "react-scripts": "0.8.5"
   },
   "devDependencies": {
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
     "electron": "1.4.12",
-    "rimraf": "^2.5.4",
     "electron-winstaller": "^2.5.1",
     "gulp": "^3.9.1",
     "gulp-bower": "0.0.13",
@@ -66,7 +69,8 @@
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^3.0.0",
     "gulp-util": "^3.0.6",
-    "main-bower-files": "^2.9.0"
+    "main-bower-files": "^2.9.0",
+    "rimraf": "^2.5.4"
   },
   "build": {
     "appId": "oa.autographa.lite",
