@@ -1,10 +1,11 @@
-const React = require('react')
-const ReactDOM = require('react-dom')
-const Constant = require("../util/constants");
-const Tabs = require('react-bootstrap/lib/Tabs');
-const Tab = require('react-bootstrap/lib/Tab');
-const session = require('electron').remote.session;
-const { dialog } = require('electron').remote;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Constant from "../util/constants";
+import Tabs from 'react-bootstrap/lib/Tabs';
+import Tab from 'react-bootstrap/lib/Tab';
+import session from 'electron';
+import { dialog } from 'electron';
+import { remote } from 'electron';
 
 
 class BookList extends React.Component {
@@ -94,8 +95,8 @@ var ChapterList = function(props) {
 	return ( <ul id="chaptersList"> { ChaptersList } </ul>)   
 }
 
-onItemClick = function(item, e) {  
+function onItemClick(item, e) {  
   global.book = item;
 }
 	
-module.exports = BookList
+export default BookList
